+++ app/controllers/survey_controller.rb
@@ -7,6 +7,7 @@ class SurveyController < ApplicationController
   before_action :find_public_design_or_redirect, only: [:new, :edit, :create, :update]
   before_action :find_or_create_subject, only: [:create]
   before_action :find_sheet_or_redirect, only: [:edit, :update]
+  before_action :redirect_on_auto_locked_sheet, only: [:edit, :update]
 
   layout 'layouts/minimal_layout'
 
@@ -67,6 +68,12 @@ class SurveyController < ApplicationController
   def redirect_without_sheet
     return if @sheet
     flash[:alert] = 'This survey no longer exists.'
+    empty_response_or_root_path(about_survey_path(survey: @design.slug))
+  end
+
+  def redirect_on_auto_locked_sheet
+    return unless @sheet.auto_locked?
+    flash[:alert] = 'This survey has been locked.'
     empty_response_or_root_path(about_survey_path(survey: @design.slug))
   end