+++ app/mailers/user_mailer.rb
@@ -37,6 +37,16 @@ class UserMailer < ApplicationMailer
          subject: "#{sheet.subject.subject_code} Submitted #{sheet.design.name}")
   end
 
+  def sheet_unlock_request(sheet, editor, requestor)
+    setup_email
+    @sheet = sheet
+    @editor = editor
+    @requestor = requestor
+    @email_to = editor.email
+    mail(to: "#{editor.name} <#{editor.email}>",
+         subject: "#{requestor.name} Requests To Unlock a Sheet on Project #{sheet.project.name}")
+  end
+
   def survey_user_link(sheet)
     setup_email
     @sheet = sheet